+++ Changelog.md
@@ -1,11 +1,17 @@
 # PHBackup changelog
 
-## 1.5.0
+## 1.5.2
 
-* Added upgrade scripts. Since now, automatic (or near automatic) updates will be possible
-* Added changelog :)
+* Reorganized functions in files
+* Added backup function mechanism
 
 ## 1.5.1
 
 * Added subdirectories for hosts to place them inside
 * Updated upgrade mechanism
+
+## 1.5.0
+
+* Added upgrade scripts. Since now, automatic (or near automatic) updates will be possible
+* Added changelog :)
+